+++ scripts/update-files.ts
@@ -10,23 +10,25 @@ let presetFiles: string[] = glob.sync("./files/presets/*.json");
 let presetPromises: Promise<object | void>[] = [];
 for (let i = 0; i < presetFiles.length; i++) {
   presetPromises.push(
-    json.read<any>(presetFiles[i]).then((data: { [key: string]: any } | void) => {
-      if (data) {
-        for (let key of Object.keys(data)) {
-          if (
-            data[key] !== null &&
-            !validator.validate(data[key] || {}).isValid()
-          ) {
-            throw new Error(`\r\n${validator.errorString}`);
-          } else {
+    json
+      .read<any>(presetFiles[i])
+      .then((data: { [key: string]: any } | void) => {
+        if (data) {
+          for (let key of Object.keys(data)) {
+            if (
+              data[key] !== null &&
+              !validator.validate(data[key] || {}).isValid()
+            ) {
+              throw new Error(`\r\n${validator.errorString}`);
+            } else {
+            }
           }
+          console.log(presetFiles[i], "\n:::\n", data);
+          return json.write(presetFiles[i], data);
+        } else {
+          return Promise.resolve();
         }
-        console.log(presetFiles[i], "\n:::\n", data);
-        return json.write(presetFiles[i], data);
-      } else {
-        return Promise.resolve();
-      }
-    }),
+      }),
   );
 }
 Promise.all(presetPromises)